refactor: unify tool provider dialogs
